move some files into the lib directory
