+++ backend/backend/backend-src/lib/db/events.ts
@@ -6,7 +6,7 @@ import {DynamoDBClient, QueryCommand} from '@aws-sdk/client-dynamodb'
 import {marshall, unmarshall} from '@aws-sdk/util-dynamodb'
 
 import {DB} from '../../constants'
-import {tickerToPackageName} from '../../utils'
+import {tickerToPackageName} from '../utils'
 
 const DB_CLIENT = new DynamoDBClient()
